Refactor OnPointerWheelChanged for precision improvements
Updated the `scale` variable type from `float` to `double` in the `OnPointerWheelChanged` method of the `BlueprintEditor` class. Adjusted calculations for `scaleMatrix`, `X`, and `Y` to use `1.0` for clarity. Added casting of `scale` back to `float` when creating the `scaleMatrix` for compatibility.

+++ sources/Blueprints.WinUI/BlueprintEditor.cs
@@ -99,14 +99,14 @@ public sealed partial class BlueprintEditor : SKXamlCanvas, IBlueprintEditor
     {
         PointerPoint pointerPoint = e.GetCurrentPoint(this);
 
-        float scale = pointerPoint.Properties.MouseWheelDelta > 0 ? 1.1f : 0.9f;
+        double scale = pointerPoint.Properties.MouseWheelDelta > 0 ? 1.1 : 0.9;
 
         Zoom *= scale;
 
-        Matrix3x2 scaleMatrix = Matrix3x2.CreateScale(scale, pointerPoint.Position.ToVector2());
+        Matrix3x2 scaleMatrix = Matrix3x2.CreateScale((float)scale, pointerPoint.Position.ToVector2());
 
-        X = (X * scaleMatrix.M11) + (pointerPoint.Position.X * (1 - scaleMatrix.M11));
-        Y = (Y * scaleMatrix.M22) + (pointerPoint.Position.Y * (1 - scaleMatrix.M22));
+        X = (X * scaleMatrix.M11) + (pointerPoint.Position.X * (1.0 - scaleMatrix.M11));
+        Y = (Y * scaleMatrix.M22) + (pointerPoint.Position.Y * (1.0 - scaleMatrix.M22));
 
         Invalidate();
     }
